+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🔄 Changed
 - `AddedAsset`'s `extraData` property is now of type `[String: RawJSON]`
+- New icon for `pendingSend` local message state
 
 # [4.25.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.25.0)
 _December 16, 2022_